refactor code for more relevant packages name
