+++ app/src/main/java/com/github/se/gatherspot/ui/navigation/TopBarSaveCancelButton.kt
@@ -2,7 +2,6 @@
 
 package com.github.se.gatherspot.ui.navigation
 
-import androidx.compose.material3.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Clear
 import androidx.compose.material.icons.filled.Done
@@ -12,9 +11,7 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.text.style.TextOverflow
-import com.github.se.gatherspot.ui.FormListener
+import com.github.se.gatherspot.model.FormListener
 
 @OptIn(ExperimentalMaterial3Api::class)
