+++ app/src/main/java/com/github/se/gatherspot/ui/theme/Shapes.kt
@@ -1,11 +1,9 @@
 package com.github.se.gatherspot.ui.theme
 
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.DrawScope
-import com.github.se.gatherspot.interest.Interests
